third_party/WebKit/LayoutTests/http/tests/notifications/serviceworker-notificationclick-openwindow-crash.html

<!doctype html>
<html>
  <head>
    <title>Notifications: Calling openWindow() in a Service Worker should not crash.</title>
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script src="../serviceworker/resources/test-helpers.js"></script>
    <script src="resources/test-helpers.js"></script>
  </head>
  <body>
    <script>
      // Tests that calling openWindow() in a Service Worker in response to a
      // notificationclick event does not cause the browser to crash.

      async_test(function(test) {
          var scope = 'resources/scope/' + location.pathname,
              script = 'resources/instrumentation-service-worker.js',
              workerInfo = null;

          testRunner.setPermission('notifications', 'granted', location.origin, location.origin);
          getActiveServiceWorkerWithMessagePort(test, script, scope).then(function(info) {
              workerInfo = info;

              // (1) Display a persistent notification with an OPENWINDOW trigger.
              return workerInfo.registration.showNotification(scope, {
                  body: 'ACTION:OPENWINDOW'
              });
          }).then(function() {
              // (2) Simulate a click on the notification, invokes the Service Worker event.
              testRunner.simulateWebNotificationClick(scope, -1 /* action_index */);

              workerInfo.port.addEventListener('message', function(event) {
                  if (typeof event.data != 'object' || !event.data.command) {
                      assert_unreached('Invalid message from the Service Worker.');
                      return;
                  }

                  // (3) Listen for confirmation by the Service Worker that the notificationclick
                  // event has been executed successfully.
                  assert_equals(event.data.command, 'click');
                  assert_equals(event.data.notification.title, scope);

                  test.done();
              });
          }).catch(unreached_rejection(test));

      }, 'Opening a window in a Service Worker should not trigger a DCHECK.');
    </script>
  </body>
</html>
